Algumas alterações como tipo Integer e correção 'nome_categoria' => 'nomeCategoria'

+++ src/main/java/com/ayvy/api_java/business/ProdutoService.java
@@ -35,7 +35,7 @@ public class ProdutoService {
     }
 
     //"DELETE"
-    public void deletarProdutoPorId(Long id){
+    public void deletarProdutoPorId(Integer id){
         repository.deleteById(id);
     }
